ransomware.py
this is a very dangerous ransomware use it wisely, you are responsible for you doings it is for educational purposes only. It actually encrytps all the files in the documents file (you can customize that in your needs. If you want you can make it encrypt every file in the computer except windows.) and then it restarts. Then when the computer is running again the ransomware asks you to type a password that password is; %2Jg@m#sg#@&p@nt@!$%. the victim has three tries to type it correctly if it is typed wrong all the three time then the documets files are deleted (you can customize it too) also if the program is closed form the task manager or the window then the documents files will be deleted as well (you can customize it as well). But if the password is correct then the documents files (you know what I would say here) would be decrypted and the program won't run unless the user does so and the computer is safe